off_chain/pyproject.toml
[project]
name = "sustainable-food-supply-chain"
version = "0.1.0"
description = "Off-Chain module backend and frontend"
authors = [
    {name = "Catalinaqi",email = "ccequispe@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyqt5 (==5.15.2)",
    "pyotp (==2.9.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pytest (>=8.3.4,<9.0.0)"
]

[tool.poetry]
packages = [{include = "Project_Sustainable_Food_Supply_Chain", from = "off_chain"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

